requirements.txt
influxdb-client==1.26.0
async-timeout==4.0.2
aiohttp==3.8.1
aiosignal==1.2.0
pyTibber==0.21.7
configobj==5.0.6
daemonize==2.5.0
